refactor(binance): Extract shared infrastructure into separate module
Refactors the monolithic Binance implementation into maintainable, reusable modules.
Spot and margin clients now share a single implementation of reconnect/backoff,
rate-limit tracking, event deduplication, error parsing, and Binance-string parsers,
reducing duplication and establishing a foundation for margin client implementation.

- Extract shared Binance infrastructure to `binance/shared.rs`:
  - AbortOnDropStream (connection manager cleanup)
  - Rate limit tracking and HTTP 429 handling
  - Event deduplication with LRU cache (keyed on instrument + ID)
  - Exponential backoff for reconnection (1s → 30s, max 10 attempts)
  - Parsing helpers: side, order_kind, time_in_force (wire strings → rustrade enums)
  - Error classification: rate-limit, auth, API rejection, balance/order errors
  - Constants for backoff, dedup cache size, rate-limit defaults
  - ~574 lines of exchange-agnostic code ready for margin/other exchange variants
- Migrate Binance Spot implementation to `binance/spot.rs`:
  - BinanceSpot and BinanceSpotConfig (moved from mod.rs)
  - REST API integration (SpotRestApi): account_snapshot, fetch_balances, fetch_open_orders, fetch_trades, open_order, cancel_order
  - WebSocket API integration (SpotWsApi): signed account_stream with user data
  - Resilience: event dedup cache, rate-limit cooldown, reconnect with fill recovery
  - Heartbeat monitoring (30s timeout forces reconnect on silence)
  - ~3909 lines focused solely on Binance Spot transport
- Add margin client placeholder in `binance/margin.rs`:
  - Documents future BinanceMargin implementation
  - Clarifies that it will reuse shared infrastructure
  - Prepares API surface for callers (both clients implement ExecutionClient trait)
- Update `binance/mod.rs`:
  - Expose only public APIs (BinanceSpot, BinanceSpotConfig)
  - Internal modules (spot, shared, margin) hidden from lib consumers
  - Reduced from ~4442 to ~15 lines
